pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "specforge"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.11"
description = "SpecForge: Speculative Decoding Training Framework"
urls = {Homepage = "https://github.com/sgl-project/SpecForge"}
dependencies = [
    "pre-commit",
    "torch==2.9.1",
    "torchaudio==2.9.1",
    "torchvision==0.24.1",
    "transformers==4.57.1",
    "qwen-vl-utils==0.0.11",
    "datasets",
    "setuptools",
    "tqdm",
    "wandb",
    "psutil",
    "numpy",
    "accelerate",
    "pydantic",
    "sglang==0.5.9",
    "openai-harmony",
    "ninja",
    "packaging",
    "yunchang",
    "tensorboard",
]

[tool.setuptools.packages.find]
exclude = ["configs*", "scripts*", "tests*"]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "unittest"
]
fa = ["flash-attn"]

[tool.setuptools.dynamic]
version = {file = "version.txt"}
